backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
notion-client>=2.2
google-genai>=1.0
PyMuPDF>=1.24
python-pptx>=1.0
python-docx>=1.1
Pillow>=10.0
